pyproject.toml
[tool.poetry]
name = "rag_neo4j_lc"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [
    { include = "app" },
]

[tool.poetry.dependencies]
python = "^3.11"
uvicorn = "^0.23.2"
langserve = {extras = ["server"], version = ">=0.0.30"}
pydantic = "<2"
rag-neo4j = {path = "packages/rag-neo4j", develop = true}
sql-pgvector = {path = "packages/sql-pgvector", develop = true}
python-dotenv = "^1.0.0"
openai = "*"
azure-identity = "*"
html2text = "*"


[tool.poetry.group.dev.dependencies]
langchain-cli = ">=0.0.15"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"